requirements.txt
pyyaml
pytest
pytest-selenium
pytest-xdist
pytest-failed-screenshot